Split utils.py into focused query modules (#180, #181)
Distributes the monolithic utils.py into seven focused modules:
db.py (execute helpers), user_queries.py, section_queries.py,
duty_queries.py, feedback_queries.py, auth_queries.py, health.py.
Applies the _build_schedule anchor-date refactor (#181) in
duty_queries.py. Removes dead user_assignments code from the login
route. Updates all route handlers, decorators, and tests to import
from the new modules instead of utils.py.
