+++ backend/bcssm_backend/decorators.py
@@ -28,7 +28,7 @@ def require_auth(f):
 
         user_id = session.get('user_id')
         if user_id is None:
-            from backend.bcssm_backend.utils import get_user_id_by_name
+            from backend.bcssm_backend.user_queries import get_user_id_by_name
             user_id = get_user_id_by_name(user_name)
             if user_id is None:
                 return jsonify({'error': 'Authentication required'}), 401
