+++ backend/tests/test_utils.py
@@ -5,17 +5,16 @@ from datetime import datetime
 import logging
 from sqlalchemy.exc import SQLAlchemyError
 from redis.exceptions import RedisError
-from backend.bcssm_backend import create_app, utils
+from backend.bcssm_backend import create_app
+from backend.bcssm_backend import duty_queries, user_queries, section_queries
+from backend.bcssm_backend import feedback_queries, auth_queries, cache_utils
+from backend.bcssm_backend.db import execute_readonly_query as _real_execute_readonly_query
+from backend.bcssm_backend.db import execute_query as _real_execute_query
 from backend.bcssm_backend.exceptions import ValidationError, AuthenticationError
-import unittest.mock
-
-# Save reference to the real function before autouse patching replaces it
-_real_execute_readonly_query = utils.execute_readonly_query
 
 # ─── 0) Autouse fixture to push a Flask app context ─────────────────────────
 @pytest.fixture(autouse=True)
 def flask_app_context(monkeypatch):
-    # Ensure create_app() loads TestingConfig
     monkeypatch.setenv('FLASK_ENV', 'testing')
     app = create_app()
     ctx = app.app_context()
@@ -23,36 +22,33 @@ def flask_app_context(monkeypatch):
     yield
     ctx.pop()
 
-# ─── 1) Autouse fixture to stub read‐only helper only ───────────────────────
+# ─── 1) Autouse fixture to stub read-only helper ────────────────────────────
 @pytest.fixture(autouse=True)
 def mock_readonly(mocker):
-    return mocker.patch('backend.bcssm_backend.utils.execute_readonly_query')
+    return mocker.patch('backend.bcssm_backend.db.execute_readonly_query')
 
 # ─── 1.5) Mock the cache to avoid Redis connection issues ───────────────────
 @pytest.fixture(autouse=True)
 def mock_cache(monkeypatch):
-    """Mock the cache to avoid Redis connection during tests"""
     fake_cache = MagicMock()
     fake_cache.get.return_value = None
     fake_cache.set.return_value = True
     fake_cache.delete.return_value = True
     fake_cache.clear.return_value = True
 
-    monkeypatch.setattr('backend.bcssm_backend.utils.cache', fake_cache)
-    # Also patch globals.cache so the cached_result decorator's lazy import gets the mock
     monkeypatch.setattr('backend.globals.cache', fake_cache)
+    monkeypatch.setattr('backend.bcssm_backend.auth_queries.cache', fake_cache)
     return fake_cache
 
 # ─── 2) Fixture to mock db.session directly (for testing execute_query) ─────
 @pytest.fixture
 def mock_db_session():
-    with patch('backend.bcssm_backend.utils.db.session') as sess:
-        # Make sess.begin() return a context manager yielding sess itself
+    with patch('backend.bcssm_backend.db.db.session') as sess:
         sess.begin.return_value.__enter__.return_value = sess
         sess.begin.return_value.__exit__.return_value = None
         yield sess
 
-# ─── 3) Utility to control "today" in date‐sensitive code ─────────────────────
+# ─── 3) Utility to control "today" in date-sensitive code ─────────────────────
 _FAKE_DATE = datetime(2025, 6, 16).date()
 
 class FakeDatetime:
@@ -67,24 +63,17 @@ class FakeDatetime:
 
 # ─── 4) Tests for get_all_users() ────────────────────────────────────────────
 def test_get_all_users_happy_path(mock_readonly, mock_cache):
-    # Arrange: return a couple of rows
     mock_readonly.return_value = [
         ('Alice Smith','SectionA','RoleA','TeamA'),
         ('Bob Jones','SectionB','RoleB','TeamB'),
     ]
-    # Act
-    result = utils.get_all_users()
-    # Assert
+    result = user_queries.get_all_users()
     assert result == ['Alice Smith','Bob Jones']
-    
-    # Verify cache was checked first
+
     mock_cache.get.assert_called_once_with('users:all:list')
-    # Verify cache was set after DB query
     mock_cache.set.assert_called_once_with('users:all:list', ['Alice Smith','Bob Jones'], timeout=900)
-    
+
     sql = mock_readonly.call_args[0][0]
-    params = None
-    assert params is None
     expected_sql = """
     SELECT
         u.name,
@@ -105,42 +94,31 @@ def test_get_all_users_happy_path(mock_readonly, mock_cache):
     assert sql.strip() == expected_sql.strip()
 
 def test_get_all_users_cache_hit(mock_readonly, mock_cache):
-    # Arrange: cache returns data
     mock_cache.get.return_value = ['Cached Alice', 'Cached Bob']
-    
-    # Act
-    result = utils.get_all_users()
-    
-    # Assert
+
+    result = user_queries.get_all_users()
+
     assert result == ['Cached Alice', 'Cached Bob']
-    
-    # Verify cache was checked
     mock_cache.get.assert_called_once_with('users:all:list')
-    # Verify DB was NOT queried
     mock_readonly.assert_not_called()
-    # Verify cache was NOT set (already had data)
     mock_cache.set.assert_not_called()
 
 def test_get_all_users_db_failure_returns_empty_list(mock_readonly, mock_cache):
     mock_readonly.side_effect = SQLAlchemyError("DB is down")
-    result = utils.get_all_users()
+    result = user_queries.get_all_users()
     assert result == []
     assert isinstance(result, list)
 
 # ─── 5) Tests for get_user_duty() ─────────────────────────────────────────────
 def test_get_user_duty_valid_today_returns_expected(monkeypatch, mock_readonly, mock_cache):
-    # Freeze today to Wednesday (weekday==2)
-    # With corrected calculation: (2 + 1) % 7 = 3
     FakeDatetime.today_weekday = 2
-    monkeypatch.setattr(utils, 'datetime', FakeDatetime)
-    
-    # Mock the get_current_cycle_week function to return a predictable value
-    monkeypatch.setattr(utils, 'get_current_cycle_week', lambda: 0)
-    
+    monkeypatch.setattr(duty_queries, 'datetime', FakeDatetime)
+    monkeypatch.setattr(duty_queries, 'get_current_cycle_week', lambda: 0)
+
     mock_readonly.return_value = [
         ("Ivy","Minis","Team Member","Team1","Lunch Duty")
     ]
-    result = utils.get_user_duty("Ivy")
+    result = duty_queries.get_user_duty("Ivy")
     assert result == {
         "user": "Ivy",
         "section": "Minis",
@@ -148,8 +126,7 @@ def test_get_user_duty_valid_today_returns_expected(monkeypatch, mock_readonly, 
         "team": "Team1",
         "duty": "Lunch Duty"
     }
-    
-    # Verify cache operations — key is now date-based
+
     cache_key_pattern = f'user:duty:Ivy:{_FAKE_DATE}'
     mock_cache.get.assert_called_once_with(cache_key_pattern)
     mock_cache.set.assert_called_once()
@@ -160,99 +137,85 @@ def test_get_user_duty_valid_today_returns_expected(monkeypatch, mock_readonly, 
     mock_readonly.assert_called_once()
     sql, params = mock_readonly.call_args[0]
     assert isinstance(params['day'], int)
-    assert params['day'] == 3  # Corrected day calculation
+    assert params['day'] == 3
     assert params['cycle_week'] == 0
 
 def test_get_user_duty_cache_hit(monkeypatch, mock_readonly, mock_cache):
-    # Arrange: cache returns duty data
     cached_duty = {
         "user": "Ivy",
-        "section": "Minis", 
+        "section": "Minis",
         "role": "Team Member",
         "team": "Team1",
         "duty": "Cached Duty"
     }
     mock_cache.get.return_value = cached_duty
-    
-    # Wednesday (weekday==2) becomes day 3 with corrected calculation
+
     FakeDatetime.today_weekday = 2
-    monkeypatch.setattr(utils, 'datetime', FakeDatetime)
-    monkeypatch.setattr(utils, 'get_current_cycle_week', lambda: 0)
-    
-    # Act
-    result = utils.get_user_duty("Ivy")
-    
-    # Assert
+    monkeypatch.setattr(duty_queries, 'datetime', FakeDatetime)
+    monkeypatch.setattr(duty_queries, 'get_current_cycle_week', lambda: 0)
+
+    result = duty_queries.get_user_duty("Ivy")
+
     assert result == cached_duty
-    
-    # Verify cache was checked — key is now date-based
     mock_cache.get.assert_called_once_with(f'user:duty:Ivy:{_FAKE_DATE}')
-    # Verify DB was NOT queried
     mock_readonly.assert_not_called()
-    # Verify cache was NOT set (already had data)
     mock_cache.set.assert_not_called()
 
 def test_get_user_duty_not_assigned_returns_error(monkeypatch, mock_readonly, mock_cache):
     FakeDatetime.today_weekday = 0
-    monkeypatch.setattr(utils, 'datetime', FakeDatetime)
-    monkeypatch.setattr(utils, 'get_current_cycle_week', lambda: 0)
-    
+    monkeypatch.setattr(duty_queries, 'datetime', FakeDatetime)
+    monkeypatch.setattr(duty_queries, 'get_current_cycle_week', lambda: 0)
+
     mock_readonly.return_value = []
-    result = utils.get_user_duty("NoOne")
+    result = duty_queries.get_user_duty("NoOne")
     assert isinstance(result, dict)
     assert "error" in result
     assert "not found or no duty" in result["error"].lower()
-    
-    # Verify error was cached (shorter timeout)
+
     mock_cache.set.assert_called_once()
     cache_set_args = mock_cache.set.call_args
-    assert cache_set_args[1]['timeout'] == 600  # Normal timeout for this case
+    assert cache_set_args[1]['timeout'] == 600
 
 def test_get_user_duty_exception_propagates_error(monkeypatch, mock_readonly, mock_cache):
     FakeDatetime.today_weekday = 0
-    monkeypatch.setattr(utils, 'datetime', FakeDatetime)
-    monkeypatch.setattr(utils, 'get_current_cycle_week', lambda: 0)
+    monkeypatch.setattr(duty_queries, 'datetime', FakeDatetime)
+    monkeypatch.setattr(duty_queries, 'get_current_cycle_week', lambda: 0)
 
     mock_readonly.side_effect = SQLAlchemyError("Something broke")
     with pytest.raises(SQLAlchemyError, match="Something broke"):
-        utils.get_user_duty("SomeUser")
+        duty_queries.get_user_duty("SomeUser")
 
-    # Error is re-raised, so no caching occurs
     mock_cache.set.assert_not_called()
 
 def test_get_user_duty_short_row(monkeypatch, mock_readonly, mock_cache):
-    """Test get_user_duty when row has fewer than 5 columns (covers len(row) < 5 branch)"""
     FakeDatetime.today_weekday = 0
-    monkeypatch.setattr(utils, 'datetime', FakeDatetime)
-    monkeypatch.setattr(utils, 'get_current_cycle_week', lambda: 0)
+    monkeypatch.setattr(duty_queries, 'datetime', FakeDatetime)
+    monkeypatch.setattr(duty_queries, 'get_current_cycle_week', lambda: 0)
 
-    # Return a row with only 3 columns instead of 5
     mock_readonly.return_value = [('Alice', 'Section1', 'Role1')]
-    result = utils.get_user_duty("Alice")
+    result = duty_queries.get_user_duty("Alice")
     assert isinstance(result, dict)
     assert "error" in result
     assert "Unexpected data format" in result["error"]
 
 def test_execute_readonly_query_db_error():
-    """Test execute_readonly_query wraps SQLAlchemyError as DatabaseError"""
     from backend.bcssm_backend.exceptions import DatabaseError
-    with patch('backend.bcssm_backend.utils.db') as mock_db:
+    with patch('backend.bcssm_backend.db.db') as mock_db:
         mock_db.engine.connect.side_effect = SQLAlchemyError("connection failed")
         with pytest.raises(DatabaseError, match="Database error"):
             _real_execute_readonly_query("SELECT 1")
 
 
 def test_execute_readonly_query_success(caplog):
-    """Test execute_readonly_query returns rows and logs when silent=False"""
     mock_conn = MagicMock()
     mock_result = MagicMock()
     mock_result.fetchall.return_value = [("row1",)]
     mock_conn.execute.return_value = mock_result
 
-    with patch('backend.bcssm_backend.utils.db') as mock_db:
+    with patch('backend.bcssm_backend.db.db') as mock_db:
         mock_db.engine.connect.return_value.__enter__ = MagicMock(return_value=mock_conn)
         mock_db.engine.connect.return_value.__exit__ = MagicMock(return_value=False)
-        with caplog.at_level(logging.INFO, logger="backend.bcssm_backend.utils"):
+        with caplog.at_level(logging.INFO, logger="backend.bcssm_backend.db"):
             rows = _real_execute_readonly_query("SELECT 1", silent=False)
 
     assert rows == [("row1",)]
@@ -261,7 +224,7 @@ def test_execute_readonly_query_success(caplog):
 # ─── 6) Tests for get_users_by_section() ───────────────────────────────────
 def test_get_users_by_section_db_error_returns_error(mock_readonly, mock_cache):
     mock_readonly.side_effect = SQLAlchemyError("DB error on section query")
-    result = utils.get_users_by_section("Minis")
+    result = section_queries.get_users_by_section("Minis")
     assert isinstance(result, dict)
     assert "Failed to fetch users by section" in result["error"]
     assert "DB error on section query" in result["error"]
@@ -269,25 +232,24 @@ def test_get_users_by_section_db_error_returns_error(mock_readonly, mock_cache):
 # ─── 7) Tests for get_all_sections() ───────────────────────────────────────
 def test_get_all_sections_with_records(mock_readonly, mock_cache):
     mock_readonly.return_value = [("Minis",), ("Micros",), ("Majors",)]
-    result = utils.get_all_sections()
+    result = section_queries.get_all_sections()
     assert result == ["Minis", "Micros", "Majors"]
-    
-    # Verify cache operations
+
     mock_cache.get.assert_called_once_with('sections:all:list')
     mock_cache.set.assert_called_once_with('sections:all:list', ["Minis", "Micros", "Majors"], timeout=3600)
-    
+
     sql = mock_readonly.call_args[0][0]
     assert "ORDER BY display_order, name" in sql.strip()
     assert sql.strip().startswith("SELECT name")
 
 def test_get_all_sections_no_records(mock_readonly, mock_cache):
     mock_readonly.return_value = []
-    result = utils.get_all_sections()
+    result = section_queries.get_all_sections()
     assert result == []
 
 def test_get_all_sections_exception(mock_readonly, mock_cache):
     mock_readonly.side_effect = SQLAlchemyError("Database failure")
-    result = utils.get_all_sections()
+    result = section_queries.get_all_sections()
     assert isinstance(result, dict)
     assert "Failed to fetch sections" in result["error"]
     assert "Database failure" in result["error"]
@@ -295,13 +257,12 @@ def test_get_all_sections_exception(mock_readonly, mock_cache):
 # ─── 8) Tests for get_all_feedback_dates() ──────────────────────────────────
 def test_get_all_feedback_dates_with_records(mock_readonly, mock_cache):
     mock_readonly.return_value = [("2025-06-01",), ("2025-05-25",)]
-    result = utils.get_all_feedback_dates()
+    result = feedback_queries.get_all_feedback_dates()
     assert result == ["2025-06-01", "2025-05-25"]
-    
-    # Verify cache operations
+
     mock_cache.get.assert_called_once_with('feedback:dates:all')
     mock_cache.set.assert_called_once_with('feedback:dates:all', ["2025-06-01", "2025-05-25"], timeout=7200)
-    
+
     sql = mock_readonly.call_args[0][0]
     expected_sql = """
     SELECT DISTINCT date
@@ -312,28 +273,25 @@ def test_get_all_feedback_dates_with_records(mock_readonly, mock_cache):
 
 def test_get_all_feedback_dates_no_records(mock_readonly, mock_cache):
     mock_readonly.return_value = []
-    result = utils.get_all_feedback_dates()
+    result = feedback_queries.get_all_feedback_dates()
     assert result == []
 
 def test_get_all_feedback_dates_exception(mock_readonly, mock_cache):
     mock_readonly.side_effect = SQLAlchemyError("Database error")
-    result = utils.get_all_feedback_dates()
+    result = feedback_queries.get_all_feedback_dates()
     assert isinstance(result, dict)
     assert "Failed to fetch feedback dates" in result["error"]
     assert "Database error" in result["error"]
 
 # ─── 9) Tests for execute_query (transactions) ──────────────────────────────
 def test_execute_query_success_with_results(mock_db_session):
-    # Arrange
     mock_result = MagicMock()
     mock_result.returns_rows = True
     mock_result.fetchall.return_value = [("Alice",)]
     mock_db_session.execute.return_value = mock_result
 
-    # Act
-    rows = utils.execute_query("SELECT name FROM users")
+    rows = _real_execute_query("SELECT name FROM users")
 
-    # Assert
     mock_db_session.execute.assert_called_once()
     mock_db_session.begin.assert_called_once()
     assert rows == [("Alice",)]
@@ -342,7 +300,7 @@ def test_execute_query_success_no_results(mock_db_session):
     mock_result = MagicMock()
     mock_result.returns_rows = False
     mock_db_session.execute.return_value = mock_result
-    rows = utils.execute_query("UPDATE users SET name='X'")
+    rows = _real_execute_query("UPDATE users SET name='X'")
     mock_db_session.execute.assert_called_once()
     mock_db_session.begin.assert_called_once()
     assert rows is None
@@ -350,275 +308,211 @@ def test_execute_query_success_no_results(mock_db_session):
 def test_execute_query_failure_triggers_rollback(mock_db_session):
     mock_db_session.execute.side_effect = SQLAlchemyError("DB error")
     with pytest.raises(Exception, match="Database error"):
-        utils.execute_query("DELETE FROM users")
+        _real_execute_query("DELETE FROM users")
     mock_db_session.rollback.assert_called_once()
 
-# Add test for get_todays_duties to verify cycle_week usage
+# Test get_todays_duties to verify cycle_week usage
 def test_get_todays_duties_uses_cycle_week(monkeypatch, mock_readonly, mock_cache):
-    # Mock datetime and cycle week
-    # Tuesday (weekday==1) becomes day 2 with corrected calculation: (1 + 1) % 7 = 2
     FakeDatetime.today_weekday = 1  # Tuesday
-    monkeypatch.setattr(utils, 'datetime', FakeDatetime)
-    monkeypatch.setattr(utils, 'get_current_cycle_week', lambda: 1)
-    
+    monkeypatch.setattr(duty_queries, 'datetime', FakeDatetime)
+    monkeypatch.setattr(duty_queries, 'get_current_cycle_week', lambda: 1)
+
     mock_readonly.return_value = [
         (1, "Kitchen Duty", "Clean kitchen", "Team A", [{"name": "Alice", "week": "Both"}], True)
     ]
-    
-    result = utils.get_todays_duties("Alice")
-    
-    # Verify cache operations - day is now 2 instead of 1
+
+    result = duty_queries.get_todays_duties("Alice")
+
     cache_key_pattern = 'duties:today:day2:cycle1:userAlice'
     mock_cache.get.assert_called_once_with(cache_key_pattern)
     mock_cache.set.assert_called_once()
-    
-    # Verify the query was called with correct parameters - day is now 2 instead of 1
+
     sql, params = mock_readonly.call_args[0]
-    assert params['day'] == 2  # Corrected day calculation
+    assert params['day'] == 2
     assert params['cycle_week'] == 1
     assert params['user_name'] == "Alice"
-    
-    # Verify simplified query structure (no CTE)
+
     assert "WITH computed_cycle" not in sql
     assert "WHERE ds.day = :day" in sql
     assert "AND ds.cycle_week = :cycle_week" in sql
-    
+
 def test_caching_behavior(mock_readonly, mock_cache):
-    """Test that caching works correctly"""
     mock_readonly.return_value = [('Alice Smith','SectionA','RoleA','TeamA')]
-    
-    # First call should hit the database
-    result1 = utils.get_all_users()
+
+    result1 = user_queries.get_all_users()
     assert mock_readonly.call_count == 1
     assert mock_cache.get.call_count == 1
     assert mock_cache.set.call_count == 1
-    
-    # Reset mock to simulate cache hit
+
     mock_cache.reset_mock()
-    mock_cache.get.return_value = ['Alice Smith']  # Simulate cache hit
+    mock_cache.get.return_value = ['Alice Smith']
     mock_readonly.reset_mock()
-    
-    # Second call should use cache
-    result2 = utils.get_all_users()
-    
-    # Verify cache was checked but DB wasn't called
+
+    result2 = user_queries.get_all_users()
+
     assert mock_cache.get.call_count == 1
-    assert mock_readonly.call_count == 0  # No DB call
-    assert mock_cache.set.call_count == 0  # No cache set (data was already cached)
-    
+    assert mock_readonly.call_count == 0
+    assert mock_cache.set.call_count == 0
+
     assert result2 == ['Alice Smith']
 
-# Test the new get_duty_schedule optimization
+# Test the get_duty_schedule optimization
 def test_get_duty_schedule_uses_parameterized_query(mock_readonly, mock_cache):
-    """Test that get_duty_schedule uses the new parameterized approach"""
     mock_readonly.return_value = [
         (1, 0, "Kitchen Duty", "Clean kitchen", "Team A", [{"name": "Alice", "week": "Both"}])
     ]
-    
-    result = utils.get_duty_schedule()
-    
-    # Verify it's a list of schedule entries
+
+    result = duty_queries.get_duty_schedule()
+
     assert isinstance(result, list)
-    assert len(result) == 14  # 2 weeks
-    
-    # Verify cache operations
+    assert len(result) == 14
+
     mock_cache.get.assert_called_once()
     cache_key = mock_cache.get.call_args[0][0]
-    assert cache_key.startswith('duties:schedule:14day:')
+    assert cache_key == 'duties:schedule:14day:anchor'
     mock_cache.set.assert_called_once()
-    
-    # Check that the query uses ANY() for parameterization instead of dynamic SQL
+
     sql, params = mock_readonly.call_args[0]
     assert "ANY(:days)" in sql
     assert "ANY(:cycles)" in sql
     assert "days" in params
     assert "cycles" in params
-    
-    # Ensure no dynamic SQL concatenation
+
     assert isinstance(params['days'], list)
     assert isinstance(params['cycles'], list)
 
 def test_get_current_cycle_week_calculation():
-    """Verify the cycle week calculation matches the camp schedule"""
     from datetime import timedelta
-    utils._cycle_week_for_date.cache_clear()
-    anchor = utils.CYCLE_ANCHOR.date()
+    duty_queries._cycle_week_for_date.cache_clear()
+    anchor = duty_queries.CYCLE_ANCHOR.date()
 
-    assert utils._cycle_week_for_date(anchor) == 0
-    assert utils._cycle_week_for_date(anchor + timedelta(weeks=1)) == 1
-    assert utils._cycle_week_for_date(anchor + timedelta(weeks=2)) == 0
+    assert duty_queries._cycle_week_for_date(anchor) == 0
+    assert duty_queries._cycle_week_for_date(anchor + timedelta(weeks=1)) == 1
+    assert duty_queries._cycle_week_for_date(anchor + timedelta(weeks=2)) == 0
 
 # ─── Tests for cache clearing methods ────────────────────────────────────────
 
 def test_clear_user_cache(mock_cache):
-    """Test clearing user-related caches"""
-    # Mock get_all_sections to return some sections to avoid DB call
-    with patch('backend.bcssm_backend.utils.get_all_sections') as mock_get_sections:
-        mock_get_sections.return_value = ['Minis', 'Micros']
-        
-        # Act
-        utils.clear_user_cache()
-        
-        # Assert
-        deleted_keys = [c.args[0] for c in mock_cache.delete.call_args_list]
-        assert 'users:all:list' in deleted_keys
-        assert 'sections:all:list' in deleted_keys
-        assert 'sections:with_users:all_v6' in deleted_keys
+    cache_utils.clear_user_cache()
+
+    deleted_keys = [c.args[0] for c in mock_cache.delete.call_args_list]
+    assert 'users:all:list' in deleted_keys
+    assert 'sections:all:list' in deleted_keys
+    assert 'sections:with_users:all_v6' in deleted_keys
 
 def test_clear_duty_cache(mock_cache):
-    """Test clearing duty-related caches delegates to clear_group."""
-    with patch('backend.bcssm_backend.utils.clear_group') as mock_cg:
-        utils.clear_duty_cache()
+    with patch('backend.bcssm_backend.cache_utils.clear_group') as mock_cg:
+        cache_utils.clear_duty_cache()
         mock_cg.assert_called_once_with("duties")
 
 def test_clear_feedback_cache(mock_cache):
-    """Test clearing feedback-related caches"""
-    # Act
-    utils.clear_feedback_cache()
-    
-    # Assert
+    cache_utils.clear_feedback_cache()
     mock_cache.delete.assert_called_once_with('feedback:dates:all')
 
 def test_clear_all_cache(mock_cache):
-    """Test clearing all caches"""
-    # Act
-    utils.clear_all_cache()
-    
-    # Assert
+    cache_utils.clear_all_cache()
     mock_cache.clear.assert_called_once()
-    # Should not call delete for specific keys when clearing all
     mock_cache.delete.assert_not_called()
 
 def test_clear_cache_methods_handle_errors_gracefully(mock_cache, caplog):
-    """Test that cache clearing methods handle Redis errors gracefully."""
     mock_cache.delete.side_effect = RedisError("Redis connection failed")
     mock_cache.clear.side_effect = RedisError("Redis connection failed")
 
-    # None of these should raise
     with caplog.at_level(logging.WARNING):
-        utils.clear_user_cache()
-        utils.clear_feedback_cache()
-        utils.clear_all_cache()
-        utils.clear_duty_cache()
+        cache_utils.clear_user_cache()
+        cache_utils.clear_feedback_cache()
+        cache_utils.clear_all_cache()
+        cache_utils.clear_duty_cache()
 
-    # clear_all_cache still logs its own warning from utils.py
     assert "Failed to clear all caches" in caplog.text
-    # clear_group swallows Redis errors and logs from cache_utils
     assert "Redis connection failed" in caplog.text
 
 def test_clear_cache_methods_are_importable():
-    """Test that all cache clearing functions can be imported"""
-    from backend.bcssm_backend.utils import (
-        clear_user_cache, 
-        clear_duty_cache, 
-        clear_feedback_cache, 
+    from backend.bcssm_backend.cache_utils import (
+        clear_user_cache,
+        clear_duty_cache,
+        clear_feedback_cache,
         clear_all_cache
     )
-    
-    # Assert functions exist and are callable
+
     assert callable(clear_user_cache)
     assert callable(clear_duty_cache)
     assert callable(clear_feedback_cache)
     assert callable(clear_all_cache)
 
 def test_cache_clearing_integration_with_data_changes(mock_readonly, mock_cache):
-    """Test typical workflow: data change -> cache clear -> fresh data"""
-    # Arrange: Populate cache first
     mock_readonly.return_value = [('Alice Smith','SectionA','RoleA','TeamA')]
-    
-    # Act 1: Get data (populates cache)
-    result1 = utils.get_all_users()
+
+    result1 = user_queries.get_all_users()
     initial_db_calls = mock_readonly.call_count
-    
-    # Act 2: Clear cache (simulating data update)
-    with patch('backend.bcssm_backend.utils.get_all_sections') as mock_get_sections:
-        mock_get_sections.return_value = []  # Empty sections to avoid additional DB calls
-        utils.clear_user_cache()
-    
-    # Act 3: Get data again (should hit DB, not cache)
-    mock_cache.get.return_value = None  # Simulate cache miss after clear
+
+    cache_utils.clear_user_cache()
+
+    mock_cache.get.return_value = None
     mock_readonly.return_value = [('Bob Jones','SectionB','RoleB','TeamB')]
-    result2 = utils.get_all_users()
-    
-    # Assert
+    result2 = user_queries.get_all_users()
+
     assert result1 == ['Alice Smith']
     assert result2 == ['Bob Jones']
-    # Should have deleted the cache keys
     mock_cache.delete.assert_called()
-    # Should have called DB twice total (once for each get_all_users call)
     assert mock_readonly.call_count == initial_db_calls + 1
-    
+
 def test_clear_user_cache_logs_correctly(mock_cache, caplog):
-    """Test that cache clearing logs appropriate messages"""
     with caplog.at_level(logging.INFO):
-        utils.clear_user_cache()
-    
-    # Check that info message was logged
+        cache_utils.clear_user_cache()
+
     assert "Cleared user-related caches" in caplog.text
 
 def test_clear_duty_cache_logs_correctly(mock_cache, caplog):
-    """Test that duty cache clearing logs correctly"""
     with caplog.at_level(logging.INFO):
-        utils.clear_duty_cache()
-    
+        cache_utils.clear_duty_cache()
+
     assert "Cleared duty-related caches" in caplog.text
 
 def test_clear_feedback_cache_logs_correctly(mock_cache, caplog):
-    """Test that feedback cache clearing logs correctly"""
     with caplog.at_level(logging.INFO):
-        utils.clear_feedback_cache()
-    
+        cache_utils.clear_feedback_cache()
+
     assert "Cleared feedback caches" in caplog.text
 
 def test_clear_all_cache_logs_correctly(mock_cache, caplog):
-    """Test that all cache clearing logs correctly"""
     with caplog.at_level(logging.INFO):
-        utils.clear_all_cache()
-    
+        cache_utils.clear_all_cache()
+
     assert "Cleared all caches" in caplog.text
 
 # ─── Tests for cache management in realistic scenarios ──────────────────────
 
 def test_cache_clear_after_user_update_scenario(mock_readonly, mock_cache):
-    """Test realistic scenario: user gets updated, cache gets cleared"""
-    # Arrange: Initial user data in cache
     mock_readonly.return_value = [('Alice Smith','SectionA','RoleA','TeamA')]
-    initial_users = utils.get_all_users()
-    
-    # Simulate user update (this would happen in your route handler)
-    # ... user update logic would go here ...
-    
-    # Act: Clear cache after user update
-    utils.clear_user_cache()
-    
-    # Arrange: New data after update
-    mock_cache.get.return_value = None  # Cache cleared
-    mock_readonly.return_value = [('Alice Jones','SectionA','RoleA','TeamA')]  # Name changed
-    
-    # Act: Get fresh data
-    updated_users = utils.get_all_users()
-    
-    # Assert
+    initial_users = user_queries.get_all_users()
+
+    cache_utils.clear_user_cache()
+
+    mock_cache.get.return_value = None
+    mock_readonly.return_value = [('Alice Jones','SectionA','RoleA','TeamA')]
+
+    updated_users = user_queries.get_all_users()
+
     assert initial_users == ['Alice Smith']
     assert updated_users == ['Alice Jones']
     mock_cache.delete.assert_called()
 
 def test_cache_clear_after_duty_schedule_update_scenario(mock_readonly, mock_cache):
-    """Duty schedule update: clear_duty_cache delegates to clear_group."""
     mock_readonly.return_value = [
         (1, 0, "Kitchen Duty", "Clean kitchen", "Team A", [{"name": "Alice", "week": "Both"}])
     ]
-    initial_schedule = utils.get_duty_schedule()
+    initial_schedule = duty_queries.get_duty_schedule()
 
-    with patch('backend.bcssm_backend.utils.clear_group') as mock_cg:
-        utils.clear_duty_cache()
+    with patch('backend.bcssm_backend.cache_utils.clear_group') as mock_cg:
+        cache_utils.clear_duty_cache()
         mock_cg.assert_called_once_with("duties")
 
     mock_cache.get.return_value = None
     mock_readonly.return_value = [
         (1, 0, "Kitchen Duty", "Clean kitchen UPDATED", "Team A", [{"name": "Bob", "week": "Both"}])
     ]
-    updated_schedule = utils.get_duty_schedule()
+    updated_schedule = duty_queries.get_duty_schedule()
 
     assert len(initial_schedule) == 14
     assert len(updated_schedule) == 14
@@ -626,76 +520,61 @@ def test_cache_clear_after_duty_schedule_update_scenario(mock_readonly, mock_cac
 # ─── Edge case tests ─────────────────────────────────────────────────────────
 
 def test_clear_duty_cache_with_different_dates(mock_cache):
-    """Duty cache clearing is now date-independent (registry-driven SCAN)."""
     for _ in range(3):
-        with patch('backend.bcssm_backend.utils.clear_group') as mock_cg:
-            utils.clear_duty_cache()
+        with patch('backend.bcssm_backend.cache_utils.clear_group') as mock_cg:
+            cache_utils.clear_duty_cache()
             mock_cg.assert_called_once_with("duties")
 
 def test_multiple_cache_clears_in_sequence(mock_cache):
-    """All cache clear methods can be called in sequence without error."""
-    utils.clear_user_cache()
-    utils.clear_duty_cache()
-    utils.clear_feedback_cache()
-    utils.clear_all_cache()
+    cache_utils.clear_user_cache()
+    cache_utils.clear_duty_cache()
+    cache_utils.clear_feedback_cache()
+    cache_utils.clear_all_cache()
 
-    # Users/sections group have static keys; feedback has one static key
     assert mock_cache.delete.call_count >= 1
-    # Only clear_all_cache calls cache.clear()
     assert mock_cache.clear.call_count >= 1
 
 # ─── Tests for get_all_sections_with_users() ─────────────────────────────────
 
 def test_get_all_sections_with_users_happy_path(mock_readonly, mock_cache):
-    """Test get_all_sections_with_users with normal data"""
-    # Arrange: return section data with users (now includes week field)
     mock_readonly.return_value = [
         ('Minis', 1, 'Alice Smith', 'Section Leader', 'Both'),
         ('Minis', 1, 'Bob Jones', 'Team Member', 'Week A'),
         ('Micros', 2, 'Charlie Brown', 'Section Leader', 'Week B'),
         ('Unassigned', 999, 'Dave Wilson', 'Team Member', 'Both')
     ]
-    
-    # Act
-    result = utils.get_all_sections_with_users()
-    
-    # Assert
+
+    result = section_queries.get_all_sections_with_users()
+
     assert isinstance(result, list)
-    assert len(result) == 3  # Minis, Micros, Unassigned
-    
-    # Check Minis section
+    assert len(result) == 3
+
     minis_section = next(s for s in result if s['name'] == 'Minis')
     assert minis_section['display_order'] == 1
     assert minis_section['user_count'] == 2
     assert len(minis_section['users']) == 2
     assert minis_section['users'][0] == {'name': 'Alice Smith', 'role': 'Section Leader', 'week': 'Both'}
     assert minis_section['users'][1] == {'name': 'Bob Jones', 'role': 'Team Member', 'week': 'Week A'}
-    
-    # Check Micros section
+
     micros_section = next(s for s in result if s['name'] == 'Micros')
     assert micros_section['user_count'] == 1
     assert micros_section['users'][0] == {'name': 'Charlie Brown', 'role': 'Section Leader', 'week': 'Week B'}
-    
-    # Check Unassigned section
+
     unassigned_section = next(s for s in result if s['name'] == 'Unassigned')
     assert unassigned_section['display_order'] == 999
     assert unassigned_section['user_count'] == 1
-    
-    # Verify cache operations (updated cache key)
+
     mock_cache.get.assert_called_once_with('sections:with_users:all_v6')
     mock_cache.set.assert_called_once_with('sections:with_users:all_v6', result, timeout=1800)
-    
-    # Verify SQL query structure
+
     sql = mock_readonly.call_args[0][0]
     assert "RIGHT JOIN users u ON s.id = u.section_id" in sql
     assert "COALESCE(s.name, 'Unassigned')" in sql
     assert "WHEN u.role = 'Admin' THEN 'Section Leader'" in sql
-    assert "u.week" in sql  # New field
+    assert "u.week" in sql
     assert "ORDER BY" in sql
 
 def test_get_all_sections_with_users_cache_hit(mock_readonly, mock_cache):
-    """Test get_all_sections_with_users with cache hit"""
-    # Arrange: cache returns data
     cached_data = [
         {
             "name": "Cached Section",
@@ -705,127 +584,93 @@ def test_get_all_sections_with_users_cache_hit(mock_readonly, mock_cache):
         }
     ]
     mock_cache.get.return_value = cached_data
-    
-    # Act
-    result = utils.get_all_sections_with_users()
-    
-    # Assert
+
+    result = section_queries.get_all_sections_with_users()
+
     assert result == cached_data
-    
-    # Verify cache was checked (updated cache key)
     mock_cache.get.assert_called_once_with('sections:with_users:all_v6')
-    # Verify DB was NOT queried
     mock_readonly.assert_not_called()
-    # Verify cache was NOT set (already had data)
     mock_cache.set.assert_not_called()
 
 def test_get_all_sections_with_users_empty_result(mock_readonly, mock_cache):
-    """Test get_all_sections_with_users with no data"""
-    # Arrange: empty result from database
     mock_readonly.return_value = []
-    
-    # Act
-    result = utils.get_all_sections_with_users()
-    
-    # Assert
+
+    result = section_queries.get_all_sections_with_users()
+
     assert result == []
-    
-    # Verify cache operations (updated cache key)
     mock_cache.get.assert_called_once_with('sections:with_users:all_v6')
     mock_cache.set.assert_called_once_with('sections:with_users:all_v6', [], timeout=1800)
 
 def test_get_all_sections_with_users_db_failure_returns_error(mock_readonly, mock_cache):
-    """Test get_all_sections_with_users with database failure"""
-    # Arrange: database error
     mock_readonly.side_effect = SQLAlchemyError("Database connection failed")
-    
-    # Act
-    result = utils.get_all_sections_with_users()
-    
-    # Assert
+
+    result = section_queries.get_all_sections_with_users()
+
     assert isinstance(result, dict)
     assert "error" in result
     assert "Failed to fetch sections with users" in result["error"]
     assert "Database connection failed" in result["error"]
-    
-    # Verify error was cached with short timeout (updated cache key)
+
     mock_cache.set.assert_called_once_with('sections:with_users:all_v6', result, timeout=60)
 
 def test_get_all_sections_with_users_admin_role_conversion(mock_readonly, mock_cache):
-    """Test that Admin role is converted to Section Leader"""
-    # Arrange: data with Admin role (now includes week field)
     mock_readonly.return_value = [
-        ('TestSection', 1, 'Admin User', 'Section Leader', 'Both'),  # Already converted in SQL
+        ('TestSection', 1, 'Admin User', 'Section Leader', 'Both'),
         ('TestSection', 1, 'Regular User', 'Team Member', 'Week A')
     ]
-    
-    # Act
-    result = utils.get_all_sections_with_users()
-    
-    # Assert
+
+    result = section_queries.get_all_sections_with_users()
+
     section = result[0]
     admin_user = next(u for u in section['users'] if u['name'] == 'Admin User')
     assert admin_user['role'] == 'Section Leader'
-    assert admin_user['week'] == 'Both'  # Check week field too
+    assert admin_user['week'] == 'Both'
 
 def test_get_all_sections_with_users_sorting(mock_readonly, mock_cache):
-    """Test that sections are sorted by display_order then name"""
-    # Arrange: unsorted data (now includes week field)
     mock_readonly.return_value = [
         ('Zebra Section', 3, 'User A', 'Team Member', 'Both'),
         ('Alpha Section', 1, 'User B', 'Team Member', 'Week A'),
         ('Beta Section', 1, 'User C', 'Team Member', 'Week B'),
         ('Unassigned', 999, 'User D', 'Team Member', 'Both')
     ]
-    
-    # Act
-    result = utils.get_all_sections_with_users()
-    
-    # Assert - should be sorted by display_order, then name
+
+    result = section_queries.get_all_sections_with_users()
+
     section_names = [s['name'] for s in result]
     assert section_names == ['Alpha Section', 'Beta Section', 'Zebra Section', 'Unassigned']
 
 # ─── Tests for get_section_statistics() ──────────────────────────────────────
 
 def test_get_section_statistics_happy_path(mock_readonly, mock_cache):
-    """Test get_section_statistics with normal data"""
-    # Arrange: return statistics data
     mock_readonly.return_value = [
-        ('Minis', 1, 5, 1, 2, 2),      # 5 total, 1 leader, 2 team leaders, 2 others
-        ('Micros', 2, 3, 1, 1, 1),     # 3 total, 1 leader, 1 team leader, 1 other
-        ('Unassigned', 999, 2, 0, 0, 2) # 2 total, 0 leaders, 0 team leaders, 2 others
+        ('Minis', 1, 5, 1, 2, 2),
+        ('Micros', 2, 3, 1, 1, 1),
+        ('Unassigned', 999, 2, 0, 0, 2)
     ]
-    
-    # Act
-    result = utils.get_section_statistics()
-    
-    # Assert
+
+    result = section_queries.get_section_statistics()
+
     assert isinstance(result, list)
     assert len(result) == 3
-    
-    # Check Minis statistics
+
     minis_stats = next(s for s in result if s['section_name'] == 'Minis')
     assert minis_stats['display_order'] == 1
     assert minis_stats['total_users'] == 5
     assert minis_stats['section_leaders'] == 1
     assert minis_stats['team_leaders'] == 2
     assert minis_stats['other_roles'] == 2
-    
-    # Check Micros statistics
+
     micros_stats = next(s for s in result if s['section_name'] == 'Micros')
     assert micros_stats['total_users'] == 3
     assert micros_stats['section_leaders'] == 1
-    
-    # Check Unassigned statistics
+
     unassigned_stats = next(s for s in result if s['section_name'] == 'Unassigned')
     assert unassigned_stats['display_order'] == 999
     assert unassigned_stats['section_leaders'] == 0
-    
-    # Verify cache operations
+
     mock_cache.get.assert_called_once_with('sections:statistics:summary')
     mock_cache.set.assert_called_once_with('sections:statistics:summary', result, timeout=3600)
-    
-    # Verify SQL query structure
+
     sql = mock_readonly.call_args[0][0]
     assert "RIGHT JOIN users u ON s.id = u.section_id" in sql
     assert "COUNT(u.id) AS total_users" in sql
@@ -834,8 +679,6 @@ def test_get_section_statistics_happy_path(mock_readonly, mock_cache):
     assert "GROUP BY s.id, s.name, s.display_order" in sql
 
 def test_get_section_statistics_cache_hit(mock_readonly, mock_cache):
-    """Test get_section_statistics with cache hit"""
-    # Arrange: cache returns data
     cached_stats = [
         {
             "section_name": "Cached Section",
@@ -847,70 +690,48 @@ def test_get_section_statistics_cache_hit(mock_readonly, mock_cache):
         }
     ]
     mock_cache.get.return_value = cached_stats
-    
-    # Act
-    result = utils.get_section_statistics()
-    
-    # Assert
+
+    result = section_queries.get_section_statistics()
+
     assert result == cached_stats
-    
-    # Verify cache was checked
     mock_cache.get.assert_called_once_with('sections:statistics:summary')
-    # Verify DB was NOT queried
     mock_readonly.assert_not_called()
-    # Verify cache was NOT set (already had data)
     mock_cache.set.assert_not_called()
 
 def test_get_section_statistics_empty_result(mock_readonly, mock_cache):
-    """Test get_section_statistics with no data"""
-    # Arrange: empty result from database
     mock_readonly.return_value = []
-    
-    # Act
-    result = utils.get_section_statistics()
-    
-    # Assert
+
+    result = section_queries.get_section_statistics()
+
     assert result == []
-    
-    # Verify cache operations
     mock_cache.get.assert_called_once_with('sections:statistics:summary')
     mock_cache.set.assert_called_once_with('sections:statistics:summary', [], timeout=3600)
 
 def test_get_section_statistics_db_failure_returns_error(mock_readonly, mock_cache):
-    """Test get_section_statistics with database failure"""
-    # Arrange: database error
     mock_readonly.side_effect = SQLAlchemyError("Statistics query failed")
-    
-    # Act
-    result = utils.get_section_statistics()
-    
-    # Assert
+
+    result = section_queries.get_section_statistics()
+
     assert isinstance(result, dict)
     assert "error" in result
     assert "Failed to fetch section statistics" in result["error"]
     assert "Statistics query failed" in result["error"]
-    
-    # Verify error was cached with short timeout
+
     mock_cache.set.assert_called_once_with('sections:statistics:summary', result, timeout=60)
 
 def test_get_section_statistics_role_counting(mock_readonly, mock_cache):
-    """Test that role counting logic works correctly"""
-    # Arrange: data with various roles including Admin
     mock_readonly.return_value = [
-        ('TestSection', 1, 6, 2, 1, 3)  # 6 total, 2 section leaders (including Admin), 1 team leader, 3 others
+        ('TestSection', 1, 6, 2, 1, 3)
     ]
-    
-    # Act
-    result = utils.get_section_statistics()
-    
-    # Assert
+
+    result = section_queries.get_section_statistics()
+
     stats = result[0]
     assert stats['total_users'] == 6
-    assert stats['section_leaders'] == 2  # Should include both Admin and Section Leader roles
+    assert stats['section_leaders'] == 2
     assert stats['team_leaders'] == 1
     assert stats['other_roles'] == 3
-    
-    # Verify totals add up
+
     assert stats['section_leaders'] + stats['team_leaders'] + stats['other_roles'] == stats['total_users']
 
 # ─── Tests for get_users_by_section() (consolidated) ─────────────────────────
@@ -922,7 +743,7 @@ def test_get_users_by_section_normal_section(mock_readonly, mock_cache):
         ('Charlie Brown', 'Team Leader')
     ]
 
-    result = utils.get_users_by_section("Minis")
+    result = section_queries.get_users_by_section("Minis")
 
     assert isinstance(result, list)
     assert result == [
@@ -944,7 +765,7 @@ def test_get_users_by_section_unassigned(mock_readonly, mock_cache):
         ('Eve Davis', 'Section Leader'),
     ]
 
-    result = utils.get_users_by_section("Unassigned")
+    result = section_queries.get_users_by_section("Unassigned")
 
     assert result == [
         {"name": "Dave Wilson", "role": "Team Member"},
@@ -963,7 +784,7 @@ def test_get_users_by_section_admin_role_mapped(mock_readonly, mock_cache):
         ('Regular User', 'Team Member'),
     ]
 
-    result = utils.get_users_by_section("TestSection")
+    result = section_queries.get_users_by_section("TestSection")
 
     admin_user = next(u for u in result if u['name'] == 'Admin User')
     assert admin_user['role'] == 'Section Leader'
@@ -972,7 +793,7 @@ def test_get_users_by_section_cache_hit(mock_readonly, mock_cache):
     cached_users = [{"name": "Cached User", "role": "Cached Role"}]
     mock_cache.get.return_value = cached_users
 
-    result = utils.get_users_by_section("TestSection")
+    result = section_queries.get_users_by_section("TestSection")
 
     assert result == cached_users
     mock_cache.get.assert_called_once_with('users:section:TestSection')
@@ -982,7 +803,7 @@ def test_get_users_by_section_cache_hit(mock_readonly, mock_cache):
 def test_get_users_by_section_empty_result(mock_readonly, mock_cache):
     mock_readonly.return_value = []
 
-    result = utils.get_users_by_section("EmptySection")
+    result = section_queries.get_users_by_section("EmptySection")
 
     assert result == []
     mock_cache.set.assert_called_once_with('users:section:EmptySection', [], timeout=1800)
@@ -990,7 +811,7 @@ def test_get_users_by_section_empty_result(mock_readonly, mock_cache):
 def test_get_users_by_section_db_failure_returns_error(mock_readonly, mock_cache):
     mock_readonly.side_effect = SQLAlchemyError("Section query failed")
 
-    result = utils.get_users_by_section("TestSection")
+    result = section_queries.get_users_by_section("TestSection")
 
     assert isinstance(result, dict)
     assert "Failed to fetch users by section" in result["error"]
@@ -1000,28 +821,24 @@ def test_get_users_by_section_db_failure_returns_error(mock_readonly, mock_cache
 # ─── Tests for clear_user_cache() updates ────────────────────────────────────
 
 def test_clear_user_cache_includes_new_keys(mock_cache):
-    """clear_user_cache clears both 'users' and 'sections' groups."""
     groups_cleared = []
     with patch(
-        'backend.bcssm_backend.utils.clear_group',
+        'backend.bcssm_backend.cache_utils.clear_group',
         side_effect=lambda g: groups_cleared.append(g),
     ):
-        utils.clear_user_cache()
+        cache_utils.clear_user_cache()
     assert set(groups_cleared) == {"users", "sections"}
 
 
 def test_clear_user_cache_handles_get_sections_error(mock_cache, caplog):
-    """clear_user_cache no longer calls get_all_sections."""
     with caplog.at_level(logging.INFO):
-        utils.clear_user_cache()
+        cache_utils.clear_user_cache()
     assert "Cleared user-related caches" in caplog.text
 
 
 def test_clear_user_cache_handles_empty_sections_list(mock_cache):
-    """clear_user_cache clears registry static keys for users+sections."""
-    utils.clear_user_cache()
+    cache_utils.clear_user_cache()
     deleted_keys = {c.args[0] for c in mock_cache.delete.call_args_list}
-    # Registry static keys in users + sections groups
     assert "users:all:list" in deleted_keys
     assert "sections:all:list" in deleted_keys
     assert "sections:with_users:all_v6" in deleted_keys
@@ -1030,69 +847,55 @@ def test_clear_user_cache_handles_empty_sections_list(mock_cache):
 # ─── Integration tests for the new methods ───────────────────────────────────
 
 def test_section_methods_integration_workflow(mock_readonly, mock_cache):
-    """Test a realistic workflow using all three new methods"""
-    # Arrange: Setup data for all three methods
-    
-    # Data for get_all_sections_with_users (now includes week field)
     sections_with_users_data = [
         ('Minis', 1, 'Alice Smith', 'Section Leader', 'Both'),
         ('Minis', 1, 'Bob Jones', 'Team Member', 'Week A'),
         ('Micros', 2, 'Charlie Brown', 'Section Leader', 'Week B')
     ]
-    
-    # Data for get_section_statistics  
+
     statistics_data = [
         ('Minis', 1, 2, 1, 0, 1),
         ('Micros', 2, 1, 1, 0, 0)
     ]
-    
-    # Data for get_users_by_section
+
     minis_users_data = [
         ('Alice Smith', 'Section Leader'),
         ('Bob Jones', 'Team Member')
     ]
 
-    # Act & Assert: Call each method and verify results
-
-    # 1. Get all sections with users
     mock_readonly.return_value = sections_with_users_data
-    sections_result = utils.get_all_sections_with_users()
+    sections_result = section_queries.get_all_sections_with_users()
     assert len(sections_result) == 2
     assert sections_result[0]['name'] == 'Minis'
     assert sections_result[0]['user_count'] == 2
 
-    # 2. Get section statistics
     mock_readonly.return_value = statistics_data
-    stats_result = utils.get_section_statistics()
+    stats_result = section_queries.get_section_statistics()
     assert len(stats_result) == 2
     assert stats_result[0]['section_name'] == 'Minis'
     assert stats_result[0]['total_users'] == 2
 
-    # 3. Get users for specific section
     mock_readonly.return_value = minis_users_data
-    users_result = utils.get_users_by_section("Minis")
+    users_result = section_queries.get_users_by_section("Minis")
     assert len(users_result) == 2
     assert users_result[0]['name'] == 'Alice Smith'
-    
-    # Verify that all methods used caching
+
     assert mock_cache.get.call_count == 3
     assert mock_cache.set.call_count == 3
 
 def test_error_handling_consistency_across_methods(mock_readonly, mock_cache):
-    """Test that all three methods handle errors consistently"""
     error_message = "Database connection lost"
     mock_readonly.side_effect = SQLAlchemyError(error_message)
-    
-    # Test each method returns error dict with consistent structure
+
     methods_to_test = [
-        (utils.get_all_sections_with_users, "Failed to fetch sections with users"),
-        (utils.get_section_statistics, "Failed to fetch section statistics"),
-        (lambda: utils.get_users_by_section("TestSection"), "Failed to fetch users by section")
+        (section_queries.get_all_sections_with_users, "Failed to fetch sections with users"),
+        (section_queries.get_section_statistics, "Failed to fetch section statistics"),
+        (lambda: section_queries.get_users_by_section("TestSection"), "Failed to fetch users by section")
     ]
-    
+
     for method, expected_error_prefix in methods_to_test:
         result = method()
-        
+
         assert isinstance(result, dict)
         assert "error" in result
         assert expected_error_prefix in result["error"]
@@ -1101,31 +904,27 @@ def test_error_handling_consistency_across_methods(mock_readonly, mock_cache):
 # ─── Tests for _cycle_week_for_date() and get_current_cycle_week() ───────────
 
 def test_cycle_week_for_date_week_a_start():
-    """CYCLE_ANCHOR is the first day of Week A (cycle_week=0)"""
-    utils._cycle_week_for_date.cache_clear()
-    assert utils._cycle_week_for_date(utils.CYCLE_ANCHOR.date()) == 0
+    duty_queries._cycle_week_for_date.cache_clear()
+    assert duty_queries._cycle_week_for_date(duty_queries.CYCLE_ANCHOR.date()) == 0
 
 def test_cycle_week_for_date_week_b_start():
-    """7 days after CYCLE_ANCHOR is the first day of Week B (cycle_week=1)"""
     from datetime import timedelta
-    utils._cycle_week_for_date.cache_clear()
-    assert utils._cycle_week_for_date((utils.CYCLE_ANCHOR + timedelta(weeks=1)).date()) == 1
+    duty_queries._cycle_week_for_date.cache_clear()
+    assert duty_queries._cycle_week_for_date((duty_queries.CYCLE_ANCHOR + timedelta(weeks=1)).date()) == 1
 
 def test_cycle_week_for_date_full_camp_schedule():
-    """All 14 days of camp have correct cycle weeks"""
     from datetime import timedelta
-    utils._cycle_week_for_date.cache_clear()
-    anchor = utils.CYCLE_ANCHOR.date()
+    duty_queries._cycle_week_for_date.cache_clear()
+    anchor = duty_queries.CYCLE_ANCHOR.date()
     for i in range(14):
         d = anchor + timedelta(days=i)
         expected = 0 if i < 7 else 1
-        assert utils._cycle_week_for_date(d) == expected, f"Day {i} ({d}) should be cycle {expected}"
+        assert duty_queries._cycle_week_for_date(d) == expected, f"Day {i} ({d}) should be cycle {expected}"
 
 def test_cycle_week_for_date_edge_cases():
-    """Boundary days around week transitions"""
     from datetime import timedelta
-    utils._cycle_week_for_date.cache_clear()
-    anchor = utils.CYCLE_ANCHOR.date()
+    duty_queries._cycle_week_for_date.cache_clear()
+    anchor = duty_queries.CYCLE_ANCHOR.date()
     test_cases = [
         (anchor + timedelta(days=-1), 1, "Day before anchor"),
         (anchor + timedelta(days=0),  0, "Anchor (Week A start)"),
@@ -1137,126 +936,116 @@ def test_cycle_week_for_date_edge_cases():
         (anchor + timedelta(days=35), 1, "Week B after month boundary"),
     ]
     for d, expected, description in test_cases:
-        result = utils._cycle_week_for_date(d)
+        result = duty_queries._cycle_week_for_date(d)
         assert result == expected, f"{description} ({d}): expected {expected}, got {result}"
 
 def test_cycle_week_for_date_caching_same_date():
-    """Same date hits cache on second call"""
-    utils._cycle_week_for_date.cache_clear()
-    d = utils.CYCLE_ANCHOR.date()
-    utils._cycle_week_for_date(d)
-    utils._cycle_week_for_date(d)
-    info = utils._cycle_week_for_date.cache_info()
+    duty_queries._cycle_week_for_date.cache_clear()
+    d = duty_queries.CYCLE_ANCHOR.date()
+    duty_queries._cycle_week_for_date(d)
+    duty_queries._cycle_week_for_date(d)
+    info = duty_queries._cycle_week_for_date.cache_info()
     assert info.hits == 1
     assert info.misses == 1
 
 def test_cycle_week_for_date_different_dates_independent():
-    """Two different dates each get their own cache entry — the core daily-eviction fix"""
     from datetime import timedelta
-    utils._cycle_week_for_date.cache_clear()
-    week_a = utils.CYCLE_ANCHOR.date()
-    week_b = (utils.CYCLE_ANCHOR + timedelta(weeks=1)).date()
+    duty_queries._cycle_week_for_date.cache_clear()
+    week_a = duty_queries.CYCLE_ANCHOR.date()
+    week_b = (duty_queries.CYCLE_ANCHOR + timedelta(weeks=1)).date()
 
-    assert utils._cycle_week_for_date(week_a) == 0
-    assert utils._cycle_week_for_date(week_b) == 1
+    assert duty_queries._cycle_week_for_date(week_a) == 0
+    assert duty_queries._cycle_week_for_date(week_b) == 1
 
-    info = utils._cycle_week_for_date.cache_info()
+    info = duty_queries._cycle_week_for_date.cache_info()
     assert info.currsize == 2
     assert info.misses == 2
 
-    # Re-calling each date returns the correct cached value, not stale
-    assert utils._cycle_week_for_date(week_a) == 0
-    assert utils._cycle_week_for_date(week_b) == 1
-    assert utils._cycle_week_for_date.cache_info().hits == 2
+    assert duty_queries._cycle_week_for_date(week_a) == 0
+    assert duty_queries._cycle_week_for_date(week_b) == 1
+    assert duty_queries._cycle_week_for_date.cache_info().hits == 2
 
 def test_cycle_week_for_date_maxsize_eviction():
-    """With maxsize=2, a third distinct date evicts the oldest entry"""
     from datetime import timedelta
-    utils._cycle_week_for_date.cache_clear()
-    anchor = utils.CYCLE_ANCHOR.date()
+    duty_queries._cycle_week_for_date.cache_clear()
+    anchor = duty_queries.CYCLE_ANCHOR.date()
 
-    utils._cycle_week_for_date(anchor)
-    utils._cycle_week_for_date(anchor + timedelta(weeks=1))
-    assert utils._cycle_week_for_date.cache_info().currsize == 2
+    duty_queries._cycle_week_for_date(anchor)
+    duty_queries._cycle_week_for_date(anchor + timedelta(weeks=1))
+    assert duty_queries._cycle_week_for_date.cache_info().currsize == 2
 
-    utils._cycle_week_for_date(anchor + timedelta(weeks=2))
-    assert utils._cycle_week_for_date.cache_info().currsize == 2  # one was evicted
+    duty_queries._cycle_week_for_date(anchor + timedelta(weeks=2))
+    assert duty_queries._cycle_week_for_date.cache_info().currsize == 2
 
 def test_cycle_week_for_date_long_term_pattern():
-    """Cycle alternates correctly week over week for 8 weeks"""
     from datetime import timedelta
-    utils._cycle_week_for_date.cache_clear()
-    anchor = utils.CYCLE_ANCHOR.date()
+    duty_queries._cycle_week_for_date.cache_clear()
+    anchor = duty_queries.CYCLE_ANCHOR.date()
     for week_num in range(8):
         d = anchor + timedelta(weeks=week_num)
         expected = week_num % 2
-        assert utils._cycle_week_for_date(d) == expected, f"Week {week_num} ({d}): expected {expected}"
+        assert duty_queries._cycle_week_for_date(d) == expected, f"Week {week_num} ({d}): expected {expected}"
 
 def test_cycle_week_for_date_matches_duty_schedule_calculation():
-    """_cycle_week_for_date and get_duty_schedule use identical logic"""
     from datetime import timedelta
-    utils._cycle_week_for_date.cache_clear()
-    anchor = utils.CYCLE_ANCHOR.date()
+    duty_queries._cycle_week_for_date.cache_clear()
+    anchor = duty_queries.CYCLE_ANCHOR.date()
     test_dates = [anchor + timedelta(weeks=n) for n in range(5)]
     for d in test_dates:
         days = (d - anchor).days
         expected = (days // 7) % 2
-        assert utils._cycle_week_for_date(d) == expected, f"Mismatch for {d}"
+        assert duty_queries._cycle_week_for_date(d) == expected, f"Mismatch for {d}"
 
 def test_cycle_week_for_date_before_anchor():
-    """Dates before CYCLE_ANCHOR follow Python floor-division semantics"""
     from datetime import timedelta
-    utils._cycle_week_for_date.cache_clear()
-    anchor = utils.CYCLE_ANCHOR.date()
+    duty_queries._cycle_week_for_date.cache_clear()
+    anchor = duty_queries.CYCLE_ANCHOR.date()
     test_cases = [
-        anchor + timedelta(days=-1),   # -1 days  → (-1//7)%2 = 1
-        anchor + timedelta(days=-7),   # -7 days  → (-7//7)%2 = 1
-        anchor + timedelta(days=-14),  # -14 days → (-14//7)%2 = 0
+        anchor + timedelta(days=-1),
+        anchor + timedelta(days=-7),
+        anchor + timedelta(days=-14),
     ]
     for d in test_cases:
         days = (d - anchor).days
         expected = (days // 7) % 2
-        assert utils._cycle_week_for_date(d) == expected, f"Before-anchor mismatch for {d}"
+        assert duty_queries._cycle_week_for_date(d) == expected, f"Before-anchor mismatch for {d}"
 
 def test_cycle_week_for_date_cache_info():
-    """_cycle_week_for_date has maxsize=2"""
-    utils._cycle_week_for_date.cache_clear()
-    info = utils._cycle_week_for_date.cache_info()
+    duty_queries._cycle_week_for_date.cache_clear()
+    info = duty_queries._cycle_week_for_date.cache_info()
     assert info.currsize == 0
     assert info.maxsize == 2
 
-    utils._cycle_week_for_date(utils.CYCLE_ANCHOR.date())
-    info = utils._cycle_week_for_date.cache_info()
+    duty_queries._cycle_week_for_date(duty_queries.CYCLE_ANCHOR.date())
+    info = duty_queries._cycle_week_for_date.cache_info()
     assert info.currsize == 1
     assert info.misses == 1
     assert info.hits == 0
 
 def test_get_current_cycle_week_delegates_to_date_helper(monkeypatch):
-    """get_current_cycle_week() passes today's date to _cycle_week_for_date"""
     captured = []
 
     def fake_cycle_week(d):
         captured.append(d)
         return 99
 
-    monkeypatch.setattr(utils, '_cycle_week_for_date', fake_cycle_week)
-    result = utils.get_current_cycle_week()
+    monkeypatch.setattr(duty_queries, '_cycle_week_for_date', fake_cycle_week)
+    result = duty_queries.get_current_cycle_week()
 
     assert result == 99
     assert len(captured) == 1
     assert captured[0] == datetime.now().date()
 
 def test_get_current_cycle_week_integration_with_get_user_duty(monkeypatch, mock_readonly, mock_cache):
-    """get_current_cycle_week integrates correctly with get_user_duty"""
     FakeDatetime.today_weekday = 0  # Monday
-    monkeypatch.setattr(utils, 'datetime', FakeDatetime)
-    monkeypatch.setattr(utils, 'get_current_cycle_week', lambda: 1)
+    monkeypatch.setattr(duty_queries, 'datetime', FakeDatetime)
+    monkeypatch.setattr(duty_queries, 'get_current_cycle_week', lambda: 1)
 
     mock_readonly.return_value = [
         ("Test User", "Test Section", "Leader", "Test Team", "Test Duty")
     ]
 
-    result = utils.get_user_duty("Test User")
+    result = duty_queries.get_user_duty("Test User")
 
     _sql, params = mock_readonly.call_args[0]
     assert params['cycle_week'] == 1
@@ -1270,73 +1059,72 @@ def test_get_current_cycle_week_integration_with_get_user_duty(monkeypatch, mock
 
 def test_get_feedback_by_date_success(monkeypatch):
     mock_exec = MagicMock(return_value=[("Minis", "Great job"), ("Majors", None)])
-    monkeypatch.setattr("backend.bcssm_backend.utils.execute_readonly_query", mock_exec)
-    result = utils.get_feedback_by_date("2025-06-07")
+    monkeypatch.setattr("backend.bcssm_backend.db.execute_readonly_query", mock_exec)
+    result = feedback_queries.get_feedback_by_date("2025-06-07")
     assert result == {"Minis": "Great job", "Majors": "No feedback available"}
 
 
 def test_get_feedback_by_date_exception(monkeypatch):
     mock_exec = MagicMock(side_effect=SQLAlchemyError("DB fail"))
-    monkeypatch.setattr("backend.bcssm_backend.utils.execute_readonly_query", mock_exec)
+    monkeypatch.setattr("backend.bcssm_backend.db.execute_readonly_query", mock_exec)
     with pytest.raises(SQLAlchemyError):
-        utils.get_feedback_by_date("2025-06-07")
+        feedback_queries.get_feedback_by_date("2025-06-07")
 
 
 # ─── Tests for get_user_info ─────────────────────────────────────────────────
 
 def test_get_user_info_found(monkeypatch):
     mock_exec = MagicMock(return_value=[("Alice", "Leader", "Minis")])
-    monkeypatch.setattr("backend.bcssm_backend.utils.execute_readonly_query", mock_exec)
-    info = utils.get_user_info("Alice")
+    monkeypatch.setattr("backend.bcssm_backend.db.execute_readonly_query", mock_exec)
+    info = user_queries.get_user_info("Alice")
     assert info == {"name": "Alice", "role": "Leader", "section": "Minis"}
 
 
 def test_get_user_info_not_found(monkeypatch):
     mock_exec = MagicMock(return_value=[])
-    monkeypatch.setattr("backend.bcssm_backend.utils.execute_readonly_query", mock_exec)
-    assert utils.get_user_info("Bob") is None
+    monkeypatch.setattr("backend.bcssm_backend.db.execute_readonly_query", mock_exec)
+    assert user_queries.get_user_info("Bob") is None
 
 
 def test_get_user_info_exception(monkeypatch):
     mock_exec = MagicMock(side_effect=SQLAlchemyError("Oops"))
-    monkeypatch.setattr("backend.bcssm_backend.utils.execute_readonly_query", mock_exec)
+    monkeypatch.setattr("backend.bcssm_backend.db.execute_readonly_query", mock_exec)
     with pytest.raises(SQLAlchemyError):
-        utils.get_user_info("Alice")
+        user_queries.get_user_info("Alice")
 
 
 def test_get_user_info_by_id_found(monkeypatch):
     mock_exec = MagicMock(return_value=[("Alice", "Leader", "Minis")])
-    monkeypatch.setattr("backend.bcssm_backend.utils.execute_readonly_query", mock_exec)
-    info = utils.get_user_info_by_id(42)
+    monkeypatch.setattr("backend.bcssm_backend.db.execute_readonly_query", mock_exec)
+    info = user_queries.get_user_info_by_id(42)
     assert info == {"name": "Alice", "role": "Leader", "section": "Minis"}
 
 
 def test_get_user_id_by_name_found(monkeypatch):
     mock_exec = MagicMock(return_value=[(7,)])
-    monkeypatch.setattr("backend.bcssm_backend.utils.execute_readonly_query", mock_exec)
-    assert utils.get_user_id_by_name("Alice") == 7
+    monkeypatch.setattr("backend.bcssm_backend.db.execute_readonly_query", mock_exec)
+    assert user_queries.get_user_id_by_name("Alice") == 7
 
 
 def test_get_user_id_by_name_not_found(monkeypatch):
     mock_exec = MagicMock(return_value=[])
-    monkeypatch.setattr("backend.bcssm_backend.utils.execute_readonly_query", mock_exec)
-    assert utils.get_user_id_by_name("Ghost") is None
+    monkeypatch.setattr("backend.bcssm_backend.db.execute_readonly_query", mock_exec)
+    assert user_queries.get_user_id_by_name("Ghost") is None
 
 
 def test_get_user_id_by_name_db_error(monkeypatch):
     from backend.bcssm_backend.exceptions import DatabaseError
     mock_exec = MagicMock(side_effect=DatabaseError("DB error"))
-    monkeypatch.setattr("backend.bcssm_backend.utils.execute_readonly_query", mock_exec)
-    assert utils.get_user_id_by_name("Alice") is None
+    monkeypatch.setattr("backend.bcssm_backend.db.execute_readonly_query", mock_exec)
+    assert user_queries.get_user_id_by_name("Alice") is None
 
 
 # ─── Tests for save_devos_feedback ───────────────────────────────────────────
 
 def test_save_devos_feedback_success(monkeypatch):
-    # Single INSERT...SELECT...RETURNING query: returns [(section_id,)] when section exists.
     mock_exec = MagicMock(return_value=[(5,)])
-    monkeypatch.setattr("backend.bcssm_backend.utils.execute_query", mock_exec)
-    utils.save_devos_feedback("Minis", "2025-06-07", "Great session", 1)
+    monkeypatch.setattr("backend.bcssm_backend.db.execute_query", mock_exec)
+    feedback_queries.save_devos_feedback("Minis", "2025-06-07", "Great session", 1)
     mock_exec.assert_called_once()
     call_params = mock_exec.call_args[0][1]
     assert call_params['section_name'] == 'Minis'
@@ -1346,25 +1134,23 @@ def test_save_devos_feedback_success(monkeypatch):
 
 
 def test_save_devos_feedback_section_not_found(monkeypatch):
-    # RETURNING returns [] when the SELECT subquery matches no section.
     mock_exec = MagicMock(return_value=[])
-    monkeypatch.setattr("backend.bcssm_backend.utils.execute_query", mock_exec)
+    monkeypatch.setattr("backend.bcssm_backend.db.execute_query", mock_exec)
     with pytest.raises(ValidationError, match="Section not found"):
-        utils.save_devos_feedback("Unknown", "2025-06-07", "feedback", 1)
+        feedback_queries.save_devos_feedback("Unknown", "2025-06-07", "feedback", 1)
 
 
 def test_save_devos_feedback_db_error(monkeypatch):
     mock_exec = MagicMock(side_effect=SQLAlchemyError("DB error"))
-    monkeypatch.setattr("backend.bcssm_backend.utils.execute_query", mock_exec)
+    monkeypatch.setattr("backend.bcssm_backend.db.execute_query", mock_exec)
     with pytest.raises(SQLAlchemyError):
-        utils.save_devos_feedback("Minis", "2025-06-07", "feedback", 1)
+        feedback_queries.save_devos_feedback("Minis", "2025-06-07", "feedback", 1)
 
 
 def test_save_devos_feedback_uses_single_query(monkeypatch):
-    # Verifies the TOCTOU-safe single-query design: exactly one execute_query call.
     mock_exec = MagicMock(return_value=[(7,)])
-    monkeypatch.setattr("backend.bcssm_backend.utils.execute_query", mock_exec)
-    utils.save_devos_feedback("Seniors", "2025-06-07", "Good work", 2)
+    monkeypatch.setattr("backend.bcssm_backend.db.execute_query", mock_exec)
+    feedback_queries.save_devos_feedback("Seniors", "2025-06-07", "Good work", 2)
     assert mock_exec.call_count == 1
     query_text = mock_exec.call_args[0][0]
     assert "INSERT INTO feedback" in query_text
@@ -1374,20 +1160,20 @@ def test_save_devos_feedback_uses_single_query(monkeypatch):
 
 def test_save_devos_feedback_clears_feedback_cache_on_success(monkeypatch):
     mock_exec = MagicMock(return_value=[(5,)])
-    monkeypatch.setattr("backend.bcssm_backend.utils.execute_query", mock_exec)
+    monkeypatch.setattr("backend.bcssm_backend.db.execute_query", mock_exec)
     mock_clear = MagicMock()
-    monkeypatch.setattr("backend.bcssm_backend.utils.clear_feedback_cache", mock_clear)
-    utils.save_devos_feedback("Minis", "2025-06-07", "Great session", 1)
-    mock_clear.assert_called_once()
+    monkeypatch.setattr("backend.bcssm_backend.feedback_queries.clear_group", mock_clear)
+    feedback_queries.save_devos_feedback("Minis", "2025-06-07", "Great session", 1)
+    mock_clear.assert_called_once_with("feedback")
 
 
 def test_save_devos_feedback_does_not_clear_cache_when_section_not_found(monkeypatch):
     mock_exec = MagicMock(return_value=[])
-    monkeypatch.setattr("backend.bcssm_backend.utils.execute_query", mock_exec)
+    monkeypatch.setattr("backend.bcssm_backend.db.execute_query", mock_exec)
     mock_clear = MagicMock()
-    monkeypatch.setattr("backend.bcssm_backend.utils.clear_feedback_cache", mock_clear)
+    monkeypatch.setattr("backend.bcssm_backend.feedback_queries.clear_group", mock_clear)
     with pytest.raises(ValidationError):
-        utils.save_devos_feedback("Unknown", "2025-06-07", "feedback", 1)
+        feedback_queries.save_devos_feedback("Unknown", "2025-06-07", "feedback", 1)
     mock_clear.assert_not_called()
 
 
@@ -1397,8 +1183,8 @@ _FAKE_HASH = "$2b$12$fakehashfakehashfakehashfakehashfakehashfakehashfakeha"
 
 def test_authenticate_user_success(mock_readonly, mocker):
     mock_readonly.return_value = [(1, "Alice", "Section Leader", "Minis", _FAKE_HASH)]
-    mocker.patch("backend.bcssm_backend.utils.bcrypt.checkpw", return_value=True)
-    result = utils.authenticate_user("Alice", "secret123")
+    mocker.patch("backend.bcssm_backend.auth_queries.bcrypt.checkpw", return_value=True)
+    result = auth_queries.authenticate_user("Alice", "secret123")
     assert result == {
         "id": 1,
         "name": "Alice",
@@ -1411,37 +1197,34 @@ def test_authenticate_user_success(mock_readonly, mocker):
 def test_authenticate_user_not_found_raises(mock_readonly):
     mock_readonly.return_value = []
     with pytest.raises(AuthenticationError):
-        utils.authenticate_user("Ghost", "secret123")
+        auth_queries.authenticate_user("Ghost", "secret123")
 
 
 def test_authenticate_user_no_password_hash_raises(mock_readonly):
-    """Users with NULL password_hash cannot log in."""
     mock_readonly.return_value = [(1, "Alice", "Section Leader", "Minis", None)]
     with pytest.raises(AuthenticationError):
-        utils.authenticate_user("Alice", "secret123")
+        auth_queries.authenticate_user("Alice", "secret123")
 
 
 def test_authenticate_user_wrong_password_raises(mock_readonly, mocker):
     mock_readonly.return_value = [(1, "Alice", "Section Leader", "Minis", _FAKE_HASH)]
-    mocker.patch("backend.bcssm_backend.utils.bcrypt.checkpw", return_value=False)
+    mocker.patch("backend.bcssm_backend.auth_queries.bcrypt.checkpw", return_value=False)
     with pytest.raises(AuthenticationError):
-        utils.authenticate_user("Alice", "wrongpassword")
+        auth_queries.authenticate_user("Alice", "wrongpassword")
 
 
 def test_authenticate_user_null_section_coalesced(mock_readonly, mocker):
     mock_readonly.return_value = [(1, "Alice", "Team Member", "Unassigned", _FAKE_HASH)]
-    mocker.patch("backend.bcssm_backend.utils.bcrypt.checkpw", return_value=True)
-    result = utils.authenticate_user("Alice", "secret123")
+    mocker.patch("backend.bcssm_backend.auth_queries.bcrypt.checkpw", return_value=True)
+    result = auth_queries.authenticate_user("Alice", "secret123")
     assert result["section_name"] == "Unassigned"
 
 
 def test_authenticate_user_uses_silent_query(mock_readonly, mocker, caplog):
-    """Auth lookup must pass silent=True and not emit sensitive data at INFO level."""
-    import logging
     mock_readonly.return_value = [(1, "Alice", "Section Leader", "Minis", _FAKE_HASH)]
-    mocker.patch("backend.bcssm_backend.utils.bcrypt.checkpw", return_value=True)
-    with caplog.at_level(logging.INFO, logger="backend.bcssm_backend.utils"):
-        utils.authenticate_user("Alice", "secret123")
+    mocker.patch("backend.bcssm_backend.auth_queries.bcrypt.checkpw", return_value=True)
+    with caplog.at_level(logging.INFO, logger="backend.bcssm_backend.auth_queries"):
+        auth_queries.authenticate_user("Alice", "secret123")
     assert mock_readonly.call_args.kwargs.get("silent") is True
     info_messages = [r.message for r in caplog.records if r.levelno == logging.INFO]
     assert not any("Alice" in m for m in info_messages)
@@ -1449,17 +1232,16 @@ def test_authenticate_user_uses_silent_query(mock_readonly, mocker, caplog):
 
 
 def test_authenticate_user_malformed_hash_raises(mock_readonly, mocker):
-    """A malformed stored hash must raise AuthenticationError, not propagate ValueError."""
     mock_readonly.return_value = [(1, "Alice", "Section Leader", "Minis", _FAKE_HASH)]
-    mocker.patch("backend.bcssm_backend.utils.bcrypt.checkpw", side_effect=ValueError("invalid hash"))
+    mocker.patch("backend.bcssm_backend.auth_queries.bcrypt.checkpw", side_effect=ValueError("invalid hash"))
     with pytest.raises(AuthenticationError):
-        utils.authenticate_user("Alice", "secret123")
+        auth_queries.authenticate_user("Alice", "secret123")
 
 
 def test_authenticate_user_db_error_propagates(mock_readonly):
     mock_readonly.side_effect = SQLAlchemyError("db down")
     with pytest.raises(SQLAlchemyError):
-        utils.authenticate_user("Alice", "secret123")
+        auth_queries.authenticate_user("Alice", "secret123")
 
 
 @pytest.mark.parametrize("role,expected", [
@@ -1470,8 +1252,8 @@ def test_authenticate_user_db_error_propagates(mock_readonly):
 ])
 def test_authenticate_user_can_edit_all_roles(mock_readonly, mocker, role, expected):
     mock_readonly.return_value = [(1, "Alice", role, "Minis", _FAKE_HASH)]
-    mocker.patch("backend.bcssm_backend.utils.bcrypt.checkpw", return_value=True)
-    result = utils.authenticate_user("Alice", "secret123")
+    mocker.patch("backend.bcssm_backend.auth_queries.bcrypt.checkpw", return_value=True)
+    result = auth_queries.authenticate_user("Alice", "secret123")
     assert result["can_edit_all"] is expected
 
 
@@ -1479,44 +1261,44 @@ def test_authenticate_user_can_edit_all_roles(mock_readonly, mocker, role, expec
 def test_cache_user_login_writes_correct_key_and_ttl(mock_cache):
     user_data = {"id": 1, "name": "Alice", "role": "Section Leader",
                  "section_name": "Minis", "can_edit_all": True}
-    utils.cache_user_login(user_data)
+    auth_queries.cache_user_login(user_data)
     mock_cache.set.assert_called_once_with("user:data:Alice", user_data, timeout=1800)
 
 
 def test_cache_user_login_swallows_redis_error(mock_cache):
     mock_cache.set.side_effect = RedisError("Redis down")
-    utils.cache_user_login({"name": "Alice"})  # must not raise
+    auth_queries.cache_user_login({"name": "Alice"})  # must not raise
 
 
 # ─── evict_user_login_cache ───────────────────────────────────────────────────
 def test_evict_user_login_cache_deletes_correct_key(mock_cache):
-    utils.evict_user_login_cache("Alice")
+    auth_queries.evict_user_login_cache("Alice")
     mock_cache.delete.assert_called_once_with("user:data:Alice")
 
 
 def test_evict_user_login_cache_swallows_redis_error(mock_cache):
     mock_cache.delete.side_effect = RedisError("Redis down")
-    utils.evict_user_login_cache("Alice")  # must not raise
+    auth_queries.evict_user_login_cache("Alice")  # must not raise
 
 
 # ─── get_user_role ────────────────────────────────────────────────────────────
 def test_get_user_role_found(mock_readonly):
     mock_readonly.return_value = [("Admin",)]
-    result = utils.get_user_role("Harrison")
+    result = auth_queries.get_user_role("Harrison")
     assert result == "Admin"
     assert mock_readonly.call_args.kwargs.get("silent") is True
 
 
 def test_get_user_role_not_found(mock_readonly):
     mock_readonly.return_value = []
-    result = utils.get_user_role("Ghost")
+    result = auth_queries.get_user_role("Ghost")
     assert result is None
 
 
 # ─── get_all_users_password_status ───────────────────────────────────────────
 def test_get_all_users_password_status(mock_readonly):
     mock_readonly.return_value = [("Alice", True), ("Bob", False)]
-    result = utils.get_all_users_password_status()
+    result = auth_queries.get_all_users_password_status()
     assert result == [
         {"name": "Alice", "has_password": True},
         {"name": "Bob", "has_password": False},
@@ -1529,7 +1311,7 @@ def test_set_user_password_found(mock_db_session):
     mock_result.returns_rows = True
     mock_result.fetchall.return_value = [(1,)]
     mock_db_session.execute.return_value = mock_result
-    assert utils.set_user_password("Alice", "$2b$12$hash") is True
+    assert auth_queries.set_user_password("Alice", "$2b$12$hash") is True
 
 
 def test_set_user_password_not_found(mock_db_session):
@@ -1537,20 +1319,17 @@ def test_set_user_password_not_found(mock_db_session):
     mock_result.returns_rows = True
     mock_result.fetchall.return_value = []
     mock_db_session.execute.return_value = mock_result
-    assert utils.set_user_password("Ghost", "$2b$12$hash") is False
+    assert auth_queries.set_user_password("Ghost", "$2b$12$hash") is False
 
 
 def test_set_user_password_redacts_hash_in_error_log(caplog):
-    """On DB failure, execute_query must log '<redacted>' not the real bcrypt hash."""
-    import logging
     from backend.bcssm_backend.exceptions import DatabaseError
-    _real_execute_query = utils.execute_query
-    with patch('backend.bcssm_backend.utils.db') as mock_db:
+    with patch('backend.bcssm_backend.db.db') as mock_db:
         mock_db.session.begin.return_value.__enter__ = MagicMock(return_value=mock_db.session)
         mock_db.session.begin.return_value.__exit__ = MagicMock(return_value=None)
         mock_db.session.execute.side_effect = SQLAlchemyError("db down")
-        with caplog.at_level(logging.ERROR, logger="backend.bcssm_backend.utils"):
-            with pytest.raises(DatabaseError):
+        with caplog.at_level(logging.ERROR, logger="backend.bcssm_backend.db"):
+            with pytest.raises(Exception):
                 _real_execute_query(
                     "UPDATE users SET password_hash = :hash WHERE name = :name RETURNING id",
                     {'hash': '$2b$12$realhash', 'name': 'Alice'},
